change dataype from int to long for epochs

+++ app/src/main/java/com/example/cfseeker/data/remote/model/RatingChange.kt
@@ -17,5 +17,5 @@ data class RatingChange(
     @SerializedName("rank")
     val rank: Int,
     @SerializedName("ratingUpdateTimeSeconds")
-    val ratingUpdateTimeSeconds: Int
+    val ratingUpdateTimeSeconds: Long
 )
